requirements.txt
agno
openai
groq
python-dotenv
duckduckgo-search
pypdf
lancedb
streamlit
pandas
tantivy
yfinance
pylance